+++ TicTacToe/TicTacToe/Controller/TTTMessagesViewController.cs
@@ -100,7 +100,7 @@ namespace TicTacToe
 		public override void WillDisplay (UITableView tableView, UITableViewCell cell, NSIndexPath indexPath)
 		{
 			TTTMessageTableViewCell messageCell = (TTTMessageTableViewCell)cell;
-			TTTMessage message = TTTMessageServer.SharedMessageServer.MessageAtIndex ((int)indexPath.Row);
+			TTTMessage message = TTTMessageServer.SharedMessageServer.MessageAtIndex (indexPath.Row);
 			cell.TextLabel.Text = message.Text;
 			cell.ImageView.Image = TTTProfile.SmallImageForIcon (message.Icon);
 
@@ -125,7 +125,7 @@ namespace TicTacToe
 
 		public override void RowSelected (UITableView tableView, NSIndexPath indexPath)
 		{
-			TTTMessage message = TTTMessageServer.SharedMessageServer.MessageAtIndex ((int)indexPath.Row);
+			TTTMessage message = TTTMessageServer.SharedMessageServer.MessageAtIndex (indexPath.Row);
 			TTTMessageTableViewCell cell = (TTTMessageTableViewCell)tableView.CellAt (indexPath);
 
 			if (selectedMessage == message) {
